(split)bug: 66620
- support running InMemoryLdapSerever in the Zimbra server JVM, used to be only supported in unit test JVM.

- renovated unit test

http://bugzilla.zimbra.com/show_bug.cgi?id=66620

Copied from Perforce
 Change: 342167

+++ src/java/com/zimbra/qa/unittest/prov/soap/Cleanup.java
@@ -16,9 +16,15 @@ package com.zimbra.qa.unittest.prov.soap;
 
 import com.zimbra.cs.account.Provisioning.CacheEntryType;
 import com.zimbra.cs.account.soap.SoapProvisioning;
+import com.zimbra.cs.ldap.unboundid.InMemoryLdapServer;
 
 public class Cleanup {
     static void deleteAll(String... domainNames) throws Exception {
+        
+        if (InMemoryLdapServer.isOn()) {
+            return;
+        }
+        
         com.zimbra.qa.unittest.prov.ldap.Cleanup.deleteAll(domainNames);
         
         SoapProvisioning prov = SoapProvisioning.getAdminInstance();
